+++ backend/src/modules/goods/goods.service.ts
@@ -2,5 +2,5 @@ import { db } from "../../db/migrate.js";
 import { goods } from "../../db/schema.js";
 
 export async function createGoods(dto: any) {
-  db.insert(goods).values(dto);
+  return await db.insert(goods).values(dto);
 }
