+++ frontend/src/components/auth_panels/auth_register.svelte
