+++ frontend/svelte.config.js
